Address compiler warnings (L4) and revision index names

+++ AdventOfCode/AdventOfCode.cpp
@@ -3,7 +3,7 @@
 #include "TestRunner.hpp"
 
 const std::string FileName = "day8.txt";
-constexpr unsigned int NumberOfTestRuns = 1;
+constexpr size_t NumberOfTestRuns = 1;
 
 int main (int)
 {
